requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
coverage==7.3.0
coverage-badge==1.1.0
crispy-bootstrap5==0.7
dj-database-url==2.0.0
Django==4.2.3
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.11.1
gunicorn==21.2.0
packaging==23.1
phonenumbers==8.13.17
Pillow==10.1.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
